Add forge lint + forge taint to CI; expose make lint-solidity
Per Foundry's security best practices: run forge lint and forge taint
on every PR touching linera-bridge/src/solidity/.

forge lint: 'make lint-solidity' (in linera-bridge/Makefile) lints the
hand-written bridge contracts (FungibleBridge, Microchain, MockERC20,
LightClient) and fails on any high/med-severity finding (forge lint's
default exit code is 0 even with warnings; we grep the output).
BridgeTypes.sol and WrappedFungibleTypes.sol are auto-generated BCS
serializers and stay out of scope. The CI job calls 'make
lint-solidity' rather than inlining the logic.

LightClient.sol _readU64LE: drop redundant 'uint64(i * 8)' cast in
the shift count — Solidity accepts any uint as a shift amount, and
the cast was forge lint's only finding on the file.

forge taint runs informational-only — exit semantics aren't stable
enough yet to gate CI on. We surface findings in logs and revisit
once the tool stabilises.

Renames the workflow job from forge-tests-and-simulation to
forge-checks now that the scope covers lint + taint + tests + sim.

+++ linera-bridge/src/solidity/LightClient.sol
@@ -268,7 +268,7 @@ contract LightClient {
     function _readU64LE(bytes memory data, uint256 pos) internal pure returns (uint64) {
         uint64 result = 0;
         for (uint256 i = 0; i < 8; i++) {
-            result |= uint64(uint8(data[pos + i])) << uint64(i * 8);
+            result |= uint64(uint8(data[pos + i])) << (i * 8);
         }
         return result;
     }
